+++ Sources/SwiftExtras/Extensions/Image+Platform.swift
@@ -26,4 +26,14 @@ extension Image {
         #endif
     }
 }
+
+extension PlatformImage: Identifiable {
+    public typealias ID = Int // swiftlint:disable:this type_name
+    /// The identifier of the color.
+    ///
+    /// This is a random hash value to make Color conform to Identifiable.
+    public var id: Int {
+        return UUID().hashValue
+    }
+}
 #endif